Consolidate interfaces from concrete steps to step contracts and make Placeholder steps extend from same base classes as concrete steps.
Migrates all of a steps interfaces (Mutating, Bypassing, FilteringBarrier...) to the step contract and ensure that placeholder and concrete steps share
common base classes (ScalarMapStep, FilterStep...). This ensures that all of a steps methods are now accessible from the StepContract type, without any
need to cast to concrete types. This also enforces complete alignment between the placeholder and concrete steps. After this change, steps with contract
interfaces should only implement the single contract interface. Any additional interfaces must be added to the contract itself. The only exception
currently is Configuring, which has been ommited from AddVertexStepContract, AddEdgeStepContract, AddPropertyStepContract, VertexStepContract, and
GraphStepContract. These are all steps which don't make use of with()-modulation in the reference implementations, and I'm reluctant to introduce
configuring into these placeholders unnecessarily due to the complexity of keeping GValues properly isolated inside Parameters. I consider a revisit of
Configuring as a followup task to be addressed separately from this change.

One additional change of note is the loss of the abstract RangeStepPlaceholder class. The shared class structure is unfortunately no longer feasible as
RangeGlobal and RangeLocal extend from distinct base classes and Java does not allow for multiple inheritance.